refactor(common): remove \EMS\CommonBundle\Common\Document
Use EMS\CommonBundle\Elasticsearch\Document\Document and his interface.
