+++ EMS/client-helper-bundle/src/Helper/Elasticsearch/ClientRequestRuntime.php
@@ -6,8 +6,9 @@ namespace EMS\ClientHelperBundle\Helper\Elasticsearch;
 
 use EMS\ClientHelperBundle\Exception\SingleResultException;
 use EMS\ClientHelperBundle\Helper\Search\Search;
-use EMS\CommonBundle\Common\Document;
 use EMS\CommonBundle\Common\EMSLink;
+use EMS\CommonBundle\Elasticsearch\Document\Document;
+use EMS\CommonBundle\Elasticsearch\Document\DocumentInterface;
 use EMS\CommonBundle\Elasticsearch\Response\Response;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
@@ -17,7 +18,7 @@ use Twig\Extension\RuntimeExtensionInterface;
 
 final class ClientRequestRuntime implements RuntimeExtensionInterface
 {
-    /** @var Document[] */
+    /** @var DocumentInterface[] */
     private array $documents = [];
 
     public function __construct(private readonly ClientRequestManager $manager, private readonly RequestStack $requestStack, private readonly LoggerInterface $logger)
@@ -42,17 +43,13 @@ final class ClientRequestRuntime implements RuntimeExtensionInterface
      * @param string|string[]|null $type
      * @param array<mixed>         $body
      */
-    public function searchOne(null|string|array $type, array $body, ?string $indexRegex = null): Document
+    public function searchOne(null|string|array $type, array $body, ?string $indexRegex = null): DocumentInterface
     {
-        $client = $this->manager->getDefault();
         try {
-            $result = $client->searchOne($type, $body, $indexRegex);
-            $document = new Document($result['_source']['_contenttype'], $result['_id'], $result['_source']);
+            return Document::fromArray($this->manager->getDefault()->searchOne($type, $body, $indexRegex));
         } catch (SingleResultException) {
             throw new NotFoundHttpException('Page not found');
         }
-
-        return $document;
     }
 
     /**
@@ -107,7 +104,7 @@ final class ClientRequestRuntime implements RuntimeExtensionInterface
         return ($response->getTotal() > 1) ? false : null;
     }
 
-    public function get(string $input): ?Document
+    public function get(string $input): ?DocumentInterface
     {
         $emsLink = EMSLink::fromText($input);
 
@@ -136,7 +133,7 @@ final class ClientRequestRuntime implements RuntimeExtensionInterface
             $this->logger->error(\sprintf('emsch_get filter found %d results for the ems key %s', $total, $input));
         }
 
-        $document = new Document($emsLink->getContentType(), $emsLink->getOuuid(), $result['hits']['hits'][0]['_source']);
+        $document = Document::fromArray($result['hits']['hits'][0]);
         $this->documents[$emsLink->__toString()] = $document;
 
         return $document;
