+++ EMS/common-bundle/src/Elasticsearch/Document/DocumentInterface.php
@@ -2,13 +2,19 @@
 
 namespace EMS\CommonBundle\Elasticsearch\Document;
 
+use EMS\CommonBundle\Common\EMSLink;
+
 interface DocumentInterface
 {
     public function getId(): string;
 
+    public function getOuuid(): string;
+
     public function getContentType(): string;
 
     public function getEmsId(): string;
+
+    public function getEmsLink(): EMSLink;
 
     /**
      * @return array<mixed>
